+++ modules/apps/commerce/headless/headless-commerce/headless-commerce-admin-pricing-impl/src/main/java/com/liferay/headless/commerce/admin/pricing/internal/util/v1_0/TierPriceUtil.java
@@ -22,6 +22,8 @@ import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.service.ServiceContext;
 import com.liferay.portal.kernel.util.GetterUtil;
 
+import java.math.BigDecimal;
+
 /**
  * @author Alessio Antonio Rendina
  */
@@ -37,7 +39,8 @@ public class TierPriceUtil {
 			tierPrice.getExternalReferenceCode(),
 			GetterUtil.getLong(tierPrice.getId()),
 			commercePriceEntry.getCommercePriceEntryId(), tierPrice.getPrice(),
-			tierPrice.getPromoPrice(), tierPrice.getMinimumQuantity(),
+			tierPrice.getPromoPrice(),
+			BigDecimal.valueOf(tierPrice.getMinimumQuantity()),
 			tierPrice.getPriceEntryExternalReferenceCode(), serviceContext);
 	}
